requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
PyYAML==6.0.1
Jinja2==3.1.2
selenium==3.141.0
selenium_wire==1.0.8
lxml==4.9.2
webdriver-manager==4.0.0
Unidecode==1.3.6
pytz==2024.1
urllib3==1.26.6